Adds new provisioning model file to supersede the old RPC ones
