+++ src/sentry/services/hybrid_cloud/organization_provisioning/impl.py
@@ -17,7 +17,7 @@ from sentry.services.hybrid_cloud.organization.serial import serialize_rpc_organ
 from sentry.services.hybrid_cloud.organization_actions.impl import (
     create_organization_and_member_for_monolith,
 )
-from sentry.services.hybrid_cloud.organization_provisioning import (
+from sentry.services.organization import (
     OrganizationProvisioningOptions,
     OrganizationProvisioningService,
 )
