+++ src/sentry/services/hybrid_cloud/organization_provisioning/model.py
@@ -1,24 +1,12 @@
-from typing import Any, Union
-
 import pydantic
 
+from sentry.services.organization.model import (  # noqa
+    OrganizationOptions,
+    OrganizationProvisioningOptions,
+    PostProvisionOptions,
+)
 
-class OrganizationOptions(pydantic.BaseModel):
-    name: str
-    slug: str
-    owning_user_id: int
-    create_default_team: bool = True
-    is_test = False
-
-
-class PostProvisionOptions(pydantic.BaseModel):
-    sentry_options: Union[Any, None]  # Placeholder for any sentry post-provisioning data
-    getsentry_options: Union[Any, None]  # Reserved for getsentry post-provisioning data
-
-
-class OrganizationProvisioningOptions(pydantic.BaseModel):
-    provision_options: OrganizationOptions
-    post_provision_options: PostProvisionOptions
+# TODO(Gabe): Remove this once GetSentry has been updated to use new model file in org provisioning
 
 
 class RpcOrganizationSlugReservation(pydantic.BaseModel):
